Added test
git-svn-id: https://openwms2005.svn.sourceforge.net/svnroot/openwms2005@1287 fda5f131-174c-0410-8613-e0905860106b

+++ org.openwms/trunk/org.openwms.core/org.openwms.core.service.spring/src/main/java/org/openwms/core/service/spring/RoleServiceImpl.java
@@ -55,7 +55,7 @@ public class RoleServiceImpl implements RoleService {
      * Instance of a {@link RoleDao}. <i>Autowired</i>.
      */
     @Autowired
-    protected RoleDao dao;
+    private RoleDao dao;
 
     /**
      * {@inheritDoc}
@@ -66,9 +66,13 @@ public class RoleServiceImpl implements RoleService {
             logger.debug("Nothing to remove just return.");
             return;
         }
+        Role role;
         for (Role r : roles) {
-            Role role = dao.findById(r.getId());
-            dao.remove(role);
+            role = r.isNew() ? dao.findByUniqueId(r.getName()) : dao.findById(r.getId());
+            if (role != null) {
+                dao.remove(role);
+            }
+            role = null;
         }
     }
